frontend: change literals to allow for postfix expressions like units
This is a major change of the grammar and the AST in preparation for the new syntax.
Motivation: Before, frequencies and durations were handeled differently than normal numeric literals in the grammar.
Now, they are a numeric literal with additional unit name as postfix. This allows to use frequencies and durations as plain expressions.
Additionally, numeric literals are now stored as their string representation with the intention to transform them when typing information is available (and check whether they can be represented as those types).