improve input reading for the EDM
We no longer require the input csv to exactly match the input stream
signature of the given Lola specification.
There is no longer a need to specify an additional input stream for the
time in order to do offline monitoring.
Therefore it is now possible to use the same spec for online and offline
monitoring.